Initial commit with old code
The code added here is the original code for which the code I've written
(the other commits in this repository) depend on. This is taken from the
old repository where this framework is extracted from. This commit have
the same time and code as commit fd203a0 in the previous repository.