Fix absolute time in reports and use more precise time slots
Output of demo data is not fixed.

Absolute times in the reports showed misleading numbers, e.g.

  0 0
  1 13
  2 1013
  3 2013
  4 3013
  5 4013

While it now is corrected to be

  0 1000
  1 2000
  2 3000
  3 4000
  4 5000
  5 6000

The more precise time slots also makes sure we don't have
slight offset from the actual data (e.g. no 13 ms offset).

Avoids measuring packets outside the time slot because we
now sets the start point from when we start the timer, instead
of when the application is initialized