framework/plot: Rework so a collection represents leaf nodes
Earlier we attached title and more to the test metadata, but by
having a collection as leaf node with a single test inside, we
can attach this title to this collection.

This allows us to swap this level in the tree, while still being
able to use title labels.

It also simplifies some part of the logic.